+++ dpnp/backend/extensions/lapack/potrf_batch.cpp
@@ -27,6 +27,7 @@
 
 // dpctl tensor headers
 #include "utils/memory_overlap.hpp"
+#include "utils/sycl_alloc_utils.hpp"
 #include "utils/type_utils.hpp"
 
 #include "linalg_exceptions.hpp"
@@ -113,7 +114,7 @@ static sycl::event potrf_batch_impl(sycl::queue &exec_q,
         }
         error_msg << ".";
 
-        sycl::free(scratchpad, exec_q);
+        dpctl::tensor::alloc_utils::sycl_free_noexcept(scratchpad, exec_q);
         throw LinAlgError(error_msg.str().c_str());
     } catch (mkl_lapack::exception const &e) {
         is_exception_caught = true;
@@ -148,7 +149,7 @@ static sycl::event potrf_batch_impl(sycl::queue &exec_q,
     if (is_exception_caught) // an unexpected error occurs
     {
         if (scratchpad != nullptr) {
-            sycl::free(scratchpad, exec_q);
+            dpctl::tensor::alloc_utils::sycl_free_noexcept(scratchpad, exec_q);
         }
         throw std::runtime_error(error_msg.str());
     }
@@ -156,7 +157,9 @@ static sycl::event potrf_batch_impl(sycl::queue &exec_q,
     sycl::event clean_up_event = exec_q.submit([&](sycl::handler &cgh) {
         cgh.depends_on(potrf_batch_event);
         auto ctx = exec_q.get_context();
-        cgh.host_task([ctx, scratchpad]() { sycl::free(scratchpad, ctx); });
+        cgh.host_task([ctx, scratchpad]() {
+            dpctl::tensor::alloc_utils::sycl_free_noexcept(scratchpad, ctx);
+        });
     });
     host_task_events.push_back(clean_up_event);
     return potrf_batch_event;